skipping file download if file exists

+++ vlbi_errors/mojave.py
@@ -67,7 +67,12 @@ def download_mojave_uv_fits(source, epochs=None, bands=None, download_dir=None):
             fname = mojave_uv_fits_fname(source, 'u', epoch)
             url = os.path.join(u_url, epoch, fname)
             print("Downloading file {}".format(fname))
-            urllib.urlretrieve(url, os.path.join(download_dir, fname))
+            path = os.path.join(download_dir, fname)
+            if os.path.isfile(path):
+                print("File {} does exist in {}."
+                      " Skipping...".format(fname, download_dir))
+                continue
+            urllib.urlretrieve(url, path)
 
     # Downloading (optionally) x, y & j-band data
     request = urllib2.Request(mojave_multifreq_url)
@@ -91,6 +96,11 @@ def download_mojave_uv_fits(source, epochs=None, bands=None, download_dir=None):
     for url in download_list:
         fname = os.path.split(url)[-1]
         print("Downloading file {}".format(fname))
+        path = os.path.join(download_dir, fname)
+        if os.path.isfile(path):
+            print("File {} does exist in {}."
+                  " Skipping...".format(fname, download_dir))
+            continue
         urllib.urlretrieve(url, os.path.join(download_dir, fname))
 
     # Downloading (optionally) l-band data
@@ -119,6 +129,11 @@ def download_mojave_uv_fits(source, epochs=None, bands=None, download_dir=None):
                     fname = mojave_uv_fits_fname(source, band, epoch)
                     url = os.path.join(l_url, epoch, fname)
                     print("Downloading file {}".format(fname))
+                    path = os.path.join(download_dir, fname)
+                    if os.path.isfile(path):
+                        print("File {} does exist in {}."
+                              " Skipping...".format(fname, download_dir))
+                        continue
                     urllib.urlretrieve(url, os.path.join(download_dir, fname))
